+++ src/enhance-collection/__tests__/index.js
@@ -3,16 +3,12 @@ import test from "ava"
 import enhanceCollection from ".."
 
 const collec = [
-  {
-    k: "ey",
-    l: "hi",
-  },
-  {
-    k: "ay",
-    q: "hu"
-  },
+  { k: "ey", l: "hi" },
+  { k: "ey", l: [ "a", "b" ] },
+  { k: "ay", q: "hu" },
   { k: "ei" },
   { k: "eye" },
+  { q: "uh" },
 ]
 
 test("filter by object", (t) => {
@@ -24,10 +20,8 @@ test("filter by object", (t) => {
       }
     ),
     [
-      {
-        k: "ey",
-        l: "hi",
-      },
+      { k: "ey", l: "hi" },
+      { k: "ey", l: [ "a", "b" ] },
     ],
     "should filter by object { key: string }"
   )
@@ -40,14 +34,9 @@ test("filter by object", (t) => {
       }
     ),
     [
-      {
-        k: "ey",
-        l: "hi",
-      },
-      {
-        k: "ay",
-        q: "hu"
-      },
+      { k: "ey", l: "hi" },
+      { k: "ey", l: [ "a", "b" ] },
+      { k: "ay", q: "hu" },
     ],
     "should filter by object { key: regexp }"
   )
@@ -58,13 +47,31 @@ test("filter by custom function", (t) => {
     enhanceCollection(
       collec,
       {
-        filter: (t) => t.k === "eye",
+        filter: (item) => item.k === "eye",
       }
     ),
     [
       { k: "eye" },
     ]
   )
+})
+
+test("filter by custom function (and warn if not boolean)", (t) => {
+  const messages = []
+  t.same(
+    enhanceCollection(
+      collec,
+      {
+        filter: (item) => item.k && item.k.indexOf("eye") + 1,
+      },
+      // console
+      { warn: (msg) => messages.push(msg) },
+    ),
+    [
+      { k: "eye" },
+    ]
+  )
+  t.ok(messages.length > 0)
 })
 
 test("filter by string", (t) => {
@@ -76,10 +83,8 @@ test("filter by string", (t) => {
       }
     ),
     [
-      {
-        k: "ey",
-        l: "hi",
-      }
+      { k: "ey", l: "hi" },
+      { k: "ey", l: [ "a", "b" ] },
     ],
   )
 })
@@ -89,14 +94,29 @@ test("multiple filters", (t) => {
     enhanceCollection(
       collec,
       {
-        filters: ["q", "k"]
+        filters: [ "q", "k" ],
       }
     ),
     [
+      { k: "ay", q: "hu" },
+    ]
+  )
+})
+
+test("mix filters ", (t) => {
+  t.same(
+    enhanceCollection(
+      collec,
       {
-      k: "ay",
-      q: "hu"
+        filters: [
+          { k: /y$/ },
+          "l",
+          (item) => Array.isArray(item.l),
+        ],
       }
+    ),
+    [
+      { k: "ey", l: [ "a", "b" ] },
     ]
   )
 })